docs: Add docs
This commit introduce docs page/system in order to write API docs
easily. The reason I choose docute is its simplicity. VuePress supports
prerendering, what I wanted, but it's not simple and does not have
versioning system. Docusaurus is not suitable for small-to-mid sized
library and it's so messy (docusaurus-init creates bunch of shit
includes Docker-related stuff, blog, noisy-top-page, etc...). Docsify
also looking good, but it lacks versioning and UX is not sufficient for
me.